+++ R/ewce_para.R
@@ -5,21 +5,20 @@
 #'  \link[EWCE]{generate_celltype_data}.
 #' @param list_names character vector of gene list names.
 #' @param gene_data data frame of gene list names and genes
-#' (see \link[MultiEWCE]{get_gene_list}).
+#' (see \link[HPOExplorer]{get_gene_lists}).
 #' @param list_name_column The name of the gene_data column
 #' that has the gene list names.
 #' @param gene_column The name of the gene_data column that contains the genes.
 #' @param save_dir_tmp Folder to save intermediate results files to
 #' (one file per gene list). Set to \code{NULL} to skip saving temporary files.
 #' @param cores The number of cores to run in parallel (e.g. 8) \code{int}.
-#' @param seed Used to set the seed for replicable results.
 #' @inheritParams EWCE::bootstrap_enrichment_test
 #' @returns Paths to saved results at "(save_dir)/(list_name).rds"
 #' (when \code{save_dir!=NULL}), or a nested list of results
 #' (when \code{save_dir==NULL}).
 #'
 #' @export
-#' @importFrom HPOExplorer load_phenotype_to_genes
+#' @importFrom HPOExplorer load_phenotype_to_genes get_gene_lists
 #' @importFrom stats setNames
 #' @importFrom parallel mclapply
 #' @importFrom EWCE bootstrap_enrichment_test
@@ -42,13 +41,11 @@ ewce_para <- function(ctd,
                       genelistSpecies="human",
                       sctSpecies="human",
                       save_dir_tmp = tempdir(),
-                      seed=2022,
                       cores=1,
                       verbose=FALSE) {
   # templateR:::source_all()
   # templateR:::args2vars(ewce_para)
 
-  set.seed(seed)
   if(!is.null(save_dir_tmp)){
     dir.create(save_dir_tmp, showWarnings = FALSE, recursive = TRUE)
   }
@@ -58,10 +55,9 @@ ewce_para <- function(ctd,
                      FUN=function(p){
     i <- which(list_names==p)
     message_parallel("Analysing: ",p," (",i,"/",length(list_names),")")
-    genes <- get_gene_list(list_name=p,
-                           gene_data=gene_data,
-                           list_name_column=list_name_column,
-                           gene_column=gene_column)
+    genes <- HPOExplorer::get_gene_lists(
+      phenotypes = p,
+      phenotype_to_genes = gene_data)[["Gene"]]
     tryCatch({
       results <- EWCE::bootstrap_enrichment_test(
         sct_data = ctd,
@@ -71,7 +67,7 @@ ewce_para <- function(ctd,
         annotLevel = annotLevel,
         genelistSpecies = genelistSpecies,
         sctSpecies = sctSpecies,
-        verbose = verbose)
+        verbose = verbose>1 )
       if(!is.null(save_dir_tmp)){
         save_path <- make_save_path(save_dir = save_dir_tmp,
                                     list_name = p)